adding scaffold for announcement and refacoting devise modules

+++ app/helpers/application_helper.rb
@@ -5,7 +5,7 @@ module ApplicationHelper
       if user_signed_in?
         content_tag(:li, link_to("<strong>#{t('general.welcome')} : #{current_user.name}</strong>".html_safe, 'javascript:void(0)' , :title => t('general.logged_user', :rel => 'tooltip', :data => { :placement => :bottom }) )) +
         content_tag(:li, '' ,:class => 'divider-vertical') +
-        content_tag(:li, link_to("Logout  <i class='icon-off'></i>".html_safe, destroy_user_session_path, :method => :delete))
+        content_tag(:li, link_to("Logout  <i class='icon-off'></i>".html_safe, destroy_user_session_path, :method => :delete, :confirm => t('devise_forms.logout_confirm')))
       else
         content_tag(:li, link_to("<b>#{t('general.register')}</b>  <b class='icon-arrow-up'></b>".html_safe, new_user_registration_path, :title => t('tooltips.register'), :rel => 'tooltip', :data => { :placement => :bottom })) +
         content_tag(:li, '', :class => 'divider-vertical') +
